Extract PostgresqlSetUp
