+++ goals/postgresql_started.go
@@ -9,50 +9,6 @@ import (
 	"denisdefreyne.com/x/ddenv/homebrew"
 )
 
-func init() {
-	core.RegisterGoal("postgresql", func(value interface{}) (core.Goal, error) {
-		detailsMap, ok := value.(map[interface{}]interface{})
-		if !ok {
-			return nil, fmt.Errorf("expected details map")
-		}
-
-		// Get version
-		version, ok := detailsMap["version"]
-		if !ok {
-			return nil, fmt.Errorf("expected version")
-		}
-		intVersion, ok := version.(int)
-		if !ok {
-			return nil, fmt.Errorf("expected integer version")
-		}
-
-		// Get env
-		rawEnv, ok := detailsMap["env"]
-		env := make(map[string]string)
-		if ok {
-			if typedEnv, ok := rawEnv.(map[interface{}]interface{}); !ok {
-				return nil, fmt.Errorf("expected env to be a map")
-			} else {
-				for rawKey, rawValue := range typedEnv {
-					if key, ok := rawKey.(string); ok {
-						if value, ok := rawValue.(string); ok {
-							env[key] = value
-						} else {
-							return nil, fmt.Errorf("expected env values to be strings")
-						}
-					} else {
-						return nil, fmt.Errorf("expected env keys to be strings")
-					}
-				}
-			}
-		}
-
-		g := PostgresqlStarted{Version: intVersion, Env: env}
-
-		return g, nil
-	})
-}
-
 type PostgresqlStarted struct {
 	Version int
 	Env     map[string]string
@@ -104,12 +60,6 @@ func (g PostgresqlStarted) Achieve() error {
 func (g PostgresqlStarted) PreGoals() []core.Goal {
 	return []core.Goal{
 		HomebrewPackageInstalled{PackageName: g.homebrewPackageName()},
-	}
-}
-
-func (g PostgresqlStarted) PostGoals() []core.Goal {
-	return []core.Goal{
-		PostgresqlShadowenvCreated{Version: g.Version, Env: g.Env},
 	}
 }
